不要なコードの削除
